+++ bindings/nostr-sdk-ffi/src/protocol/event/builder.rs
@@ -836,20 +836,20 @@ impl EventBuilder {
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/34.md>
     #[uniffi::constructor]
-    pub fn git_repository_announcement(data: GitRepositoryAnnouncement) -> Self {
-        Self {
-            inner: nostr::EventBuilder::git_repository_announcement(data.into()),
-        }
+    pub fn git_repository_announcement(data: GitRepositoryAnnouncement) -> Result<Self> {
+        Ok(Self {
+            inner: nostr::EventBuilder::git_repository_announcement(data.into())?,
+        })
     }
 
     /// Git Issue
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/34.md>
     #[uniffi::constructor]
-    pub fn git_issue(issue: GitIssue) -> Self {
-        Self {
-            inner: nostr::EventBuilder::git_issue(issue.into()),
-        }
+    pub fn git_issue(issue: GitIssue) -> Result<Self> {
+        Ok(Self {
+            inner: nostr::EventBuilder::git_issue(issue.into())?,
+        })
     }
 
     /// Git Patch
@@ -858,7 +858,7 @@ impl EventBuilder {
     #[uniffi::constructor]
     pub fn git_patch(patch: GitPatch) -> Result<Self> {
         Ok(Self {
-            inner: nostr::EventBuilder::git_patch(patch.try_into()?),
+            inner: nostr::EventBuilder::git_patch(patch.try_into()?)?,
         })
     }
 }